+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.13] - 2023-01-12
+- Added convert, affine!
+- Added broadcast for GeoArray, so `ga .+ 1` isa `GeoArray`
+
 ## [0.7.12] - 2023-01-12
 - Fix interpolation, update to GeoStatsSolvers
 - Fix indexing bug in non-singleton sized GeoArrays
@@ -35,7 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed iterate specification so `sum` on a GeoArray is correct
 
-[unreleased]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.9...HEAD
+[unreleased]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.13...HEAD
+[0.7.13]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.12...v0.7.13
 [0.7.12]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.11...v0.7.12
 [0.7.11]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.10...v0.7.11
 [0.7.10]: https://github.com/evetion/GeoArrays.jl/compare/v0.7.9...v0.7.10